fix: clear relics from arrays
We were previously updating relics to an object populated with empty strings when cleare, based on the index. This means that if you added 3 relics and cleared the first two, the alternativeRelics array would still be of length 3. Within the UI components, though, the array would be of length 1 so the indices wouldn't match.

This change clears relics from the array entirely so that indices always match.